added SQL Creation Script and prior/next Month button functionality
I need a way to stall the next/ prior month button so that it can't be pressed twice before the page reloads I think, if you press it quickly it's skipping past the enable/disable logic I think. Also I added a SQL script for the creation of a database. If you could look at both of those and potentially review the rest of the code that would be awesome. I don't have any functionality for the database yet I just wanted to have a good script in place to show I'm working on connecting a database. Thanks!